# box: introduce `tuple:info()` and `space:stat()`