--- a/changelogs/unreleased/gh-6762-add-tuple-info-and-space-stat.md
+++ b/changelogs/unreleased/gh-6762-add-tuple-info-and-space-stat.md
@@ -1,0 +1,4 @@
+## feature/core
+
+* Introduced new methods `tuple:info()` and `space:stat()` with the detailed
+  information on memory consumed by data (gh-6762).