--- a/src/box/memtx_engine.cc
+++ b/src/box/memtx_engine.cc
@@ -617,10 +617,7 @@ memtx_engine_commit(struct engine *engine, struct txn *txn)
 	stailq_foreach_entry(stmt, &txn->stmts, next) {
 		if (memtx_tx_manager_use_mvcc_engine) {
 			assert(stmt->space->engine == engine);
-			struct memtx_space *mspace =
-				(struct memtx_space *)stmt->space;
-			size_t *bsize = &mspace->bsize;
-			memtx_tx_history_commit_stmt(stmt, bsize);
+			memtx_tx_history_commit_stmt(stmt);
 		}
 		if (stmt->engine_savepoint != NULL) {
 			struct space *space = stmt->space;
@@ -679,7 +676,7 @@ memtx_engine_rollback_statement(struct engine *engine, struct txn *txn,
 		}
 	}
 
-	memtx_space_update_bsize(space, new_tuple, old_tuple);
+	memtx_space_update_tuple_stat(space, new_tuple, old_tuple);
 	if (old_tuple != NULL)
 		tuple_ref(old_tuple);
 	if (new_tuple != NULL)