#### `tuple_object:info()` - Get information about the tuple

Returns a table with the following fields:

 * `data_size` - Size of the MessagePack data in the tuple.
   This number equals to number returned by `tuple_object:bsize()`.
 * `header_size` - Size of the internal tuple header.
 * `field_map_size` - Size of the field map.
   Field map is used to speed up access to indexed fields of the tuple.
 * `waste_size` - The amount of excess memory used to store the tuple in mempool.
 * `arena` - Type of the arena where the tuple is allocated.
   Possible values are: "memtx", "malloc", "runtime".

---

#### `space_object:stat()` - Get statistics on the memory usage

Returns a table with the cumulative statistics on the memory usage by tuples in the space.
The statistics is grouped by arena types: "memtx" or "malloc".
For a detailed description of each field see `tuple_object:info()`.

Note: The statistics is collected only for memtx storage engine.
For other types of spaces, an empty table is returned.

Example:

```yaml
tarantool> box.space.test:stat()
---
- tuple:
    memtx:
      data_size: 5100699
      header_size: 96
      field_map_size: 40
      waste_size: 143093
    malloc:
      data_size: 18850077
      header_size: 70
      field_map_size: 28
      waste_size: 0
...
```

Closes https://github.com/tarantool/tarantool/issues/6762

**Do not merge until https://github.com/tarantool/small/pull/77 and #9251 are merged!**