@Gumix 2023-10-05
Benchmarked using:
- https://github.com/tarantool/tarantool/pull/7083

with:

```diff
--- a/perf/memtx.cc
+++ b/perf/memtx.cc
@@ -356,6 +356,32 @@ BENCHMARK_F(MemtxFixture, TreeSelectAll)
        state.SetItemsProcessed(counter);
 }

+/**
+ * Benchmark random `replace`s of existing keys in the tree index.
+ * The key subset is regenerated through every iteration.
+ */
+BENCHMARK_F(MemtxFixture, TreeReplaceRandomExistingKeys)
+(benchmark::State &state)
+{
+       auto itr = key_subset.begin();
+       int64_t counter = 0;
+       for (MAYBE_UNUSED auto _ : state) {
+               if (itr == key_subset.end()) {
+                       state.PauseTiming();
+                       generate_key_subset();
+                       state.ResumeTiming();
+                       itr = key_subset.begin();
+               }
+               struct tuple *result;
+               if (::box_replace(sid, itr->first, itr->second, &result) != 0)
+                       panic("failed to replace the tuple");
+               benchmark::DoNotOptimize(result);
+               ++counter;
+               ++itr;
+       }
+       state.SetItemsProcessed(counter);
+}
+
 BENCHMARK_MAIN();

 #include "debug_warning.h"
```

#### How to run:

```
$ taskset 0x2 ./build/perf/memtx.perftest \
  --benchmark_min_warmup_time=5 \
  --benchmark_repetitions=30 \
  --benchmark_report_aggregates_only=true \
  --benchmark_display_aggregates_only=true \
  --benchmark_filter=TreeReplaceRandomExistingKeys
```

#### Without the patch:

```
------------------------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------
MemtxFixture/TreeReplaceRandomExistingKeys_mean         1186 ns         1186 ns           30 items_per_second=844.764k/s
MemtxFixture/TreeReplaceRandomExistingKeys_median       1199 ns         1199 ns           30 items_per_second=833.869k/s
MemtxFixture/TreeReplaceRandomExistingKeys_stddev       48.0 ns         48.0 ns           30 items_per_second=35.1014k/s
MemtxFixture/TreeReplaceRandomExistingKeys_cv           4.05 %          4.05 %            30 items_per_second=4.16%
```

#### With the patch:

```
------------------------------------------------------------------------------------------------------------
Benchmark                                                  Time             CPU   Iterations UserCounters...
------------------------------------------------------------------------------------------------------------
MemtxFixture/TreeReplaceRandomExistingKeys_mean         1207 ns         1207 ns           30 items_per_second=829.972k/s
MemtxFixture/TreeReplaceRandomExistingKeys_median       1205 ns         1205 ns           30 items_per_second=829.844k/s
MemtxFixture/TreeReplaceRandomExistingKeys_stddev       56.8 ns         56.8 ns           30 items_per_second=37.5861k/s
MemtxFixture/TreeReplaceRandomExistingKeys_cv           4.70 %          4.70 %            30 items_per_second=4.53%
```